tp/back/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
PyJWT
passlib[bcrypt]
python-multipart
